+++ src/model/main/tableNode.ts
@@ -36,16 +36,16 @@ export class TableNode extends Node implements CopyAble {
 
     public async getChildren(isRresh: boolean = false): Promise<Node[]> {
         let columnNodes = DatabaseCache.getColumnListOfTable(this.id);
-        if (columnNodes && !isRresh) {
+        if (columnNodes && !isRresh && this.collapsibleState != vscode.TreeItemCollapsibleState.Expanded) {
             return columnNodes;
         }
         return QueryUnit.queryPromise<ColumnMeta[]>(await ConnectionManager.getConnection(this), `SELECT COLUMN_NAME name,DATA_TYPE simpleType,COLUMN_TYPE type,COLUMN_COMMENT comment,COLUMN_KEY \`key\`,IS_NULLABLE nullable,CHARACTER_MAXIMUM_LENGTH maxLength,COLUMN_DEFAULT defaultValue,EXTRA extra FROM information_schema.columns WHERE table_schema = '${this.database}' AND table_name = '${this.table}' ORDER BY ORDINAL_POSITION;`)
             .then((columns) => {
-                columnNodes = columns.map<ColumnNode>((column) => {
+                columnNodes = columns.map<ColumnNode>((column, index) => {
                     if (column && column.key == "PRI") {
                         MockRunner.primaryKeyMap[this.getConnectId()] = column.name
                     }
-                    return new ColumnNode(this.table, column, this.info);
+                    return new ColumnNode(this.table, column, this, index);
                 });
                 DatabaseCache.setColumnListOfTable(this.id, columnNodes);
 
@@ -120,7 +120,7 @@ ADD
     }
 
     public async openInNew() {
-        const pageSize=Global.getConfig(ConfigKey.DEFAULT_LIMIT);
+        const pageSize = Global.getConfig(ConfigKey.DEFAULT_LIMIT);
         const sql = `SELECT * FROM ${Util.wrap(this.database)}.${Util.wrap(this.table)} LIMIT ${pageSize};`;
         const connection = await ConnectionManager.getConnection(this);
         const executeTime = new Date().getTime();